simpleWebApp/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.51.0
greenlet==3.0.3
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
PyMySQL==1.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
tzdata==2024.1
Werkzeug==3.0.3
zipp==3.18.1
